requirements.txt
# Core ML
torch>=2.1.0
torchvision>=0.16.0
snntorch>=0.9.0

# Data processing
numpy
pandas
pyarrow
h5py

# Vector search
faiss-cpu

# Configuration
pyyaml

# Training utilities
tqdm
tensorboard
scikit-learn

# Visualization
matplotlib
seaborn

# Evaluation
scipy

# Dataset download
requests
